Split messages into a dedicated crate
