+++ node/ledger/Cargo.toml
@@ -48,6 +48,9 @@ version = "1"
 [dependencies.reqwest]
 version = "0.11"
 
+[dependencies.snarkos-node-messages]
+path = "../messages"
+
 [dependencies.snarkos-node-router]
 path = "../router"
