+++ node/messages/Cargo.toml
@@ -1,8 +1,8 @@
 [package]
-name = "snarkos-node"
+name = "snarkos-node-messages"
 version = "2.0.2"
 authors = [ "The Aleo Team <hello@aleo.org>" ]
-description = "A node for a decentralized operating system"
+description = "Node messages for a decentralized operating system"
 homepage = "https://aleo.org"
 repository = "https://github.com/AleoHQ/snarkOS"
 keywords = [
@@ -19,34 +19,36 @@ edition = "2021"
 [dependencies.anyhow]
 version = "1"
 
-[dependencies.async-trait]
-version = "0.1"
+[dependencies.bincode]
+version = "1.0"
 
-[dependencies.rand]
-version = "0.8"
-default-features = false
+[dependencies.bytes]
+version = "1.0.0"
 
-[dependencies.serde_json]
+[dependencies.serde]
 version = "1"
 
-[dependencies.snarkos-account]
-path = "../account"
-
 [dependencies.snarkos-node-executor]
-path = "./executor"
-
-[dependencies.snarkos-node-ledger]
-path = "./ledger"
-
-[dependencies.snarkos-node-router]
-path = "./router"
+path = "../executor"
 
 [dependencies.snarkvm]
 workspace = true
 
 [dependencies.tokio]
 version = "1.21"
-features = ["rt", "signal"]
+features = [
+  "io-util",
+  "macros",
+  "net",
+  "rt-multi-thread",
+  "signal",
+  "sync",
+  "time"
+]
+
+[dependencies.tokio-util]
+version = "0.7"
+features = [ "codec" ]
 
 [dependencies.tracing]
 version = "0.1"
