+++ node/messages/src/helpers/data.rs
